// remix-solidity/src/compiler/types.ts
export interface MessageToWorker {
  cmd: string,
  job?: number,
  input?: CompilerInput,
  data?: string
}